requirements.txt
asgiref==3.9.1
certifi==2024.12.14
charset-normalizer==3.4.1
crypto==1.4.1
Django==5.2.4
djangorestframework==3.16.0
idna==3.10
Naked==0.1.32
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
